refactor(oauth2-mock): extract portal discovery into portal-discovery.ts
refactor: split discoverPortalConfigs into helpers to reduce cognitive complexity (SonarQube S3776)
